third_party/WebKit/LayoutTests/platform/win/external/wpt/screen-orientation/lock-bad-argument-expected.txt

This is a testharness.js-based test.
PASS screen.orientation.lock() must throw given invalid input.
PASS screen.orientation.lock() must throw when the input is missing.
Harness: the test ran to completion.
